Copy the relevant parts of AsyncFunctionStart into ModuleExecution
ModuleExecution now returns a promise derived from evaluating the body kind of as if it were an async function. This duplicates a decent amount of spec from AsyncFunctionStart; we should refactor this next.

Nobody uses this promise return value yet. And, it's not grammatically allowed to use `await` yet. But, this is a start.