Propogate promise-ness to InnerModuleEvaluation
Now InnerModuleEvaluation returns a promise for the index, instead of the index itself. It also awaits recursive invocations of itself, as well as the invocation of ModuleExecution.

This implements variant A (sequential, blocking execution) since it is easier to write spec text for; we can in the future easily (?) convert to variant B by making local modifications to InnerModuleEvaluation's child-evaluation loop.