Normative: Ensure that all dependencies of async modules are triggered
This patch ensures that async parent modules are evaluated, when ready,
for both [[Async]] and non-[[Async]] modules. A sync module may have
async parent modules when it depends on an async module. For example:

a.mjs			b.mjs			c.mjs
import "./b.mjs";	import "./c.mjs";	await undefined;

In this case, b.mjs will have [[Async]] as false, but it will have
a.mjs in its [[AsyncParentModules]]. When c.mjs is done executing,
it will run b.mjs. But previously, a.mjs would never be executed.

This patch also fixes up the resolution of the top level Promise,
which would otherwise be missed in some cases (e.g., it was never
resolved for synchronous modules).