Normative: Allow export declarations to use await
This patch ensures that the [+Await] syntactic parameter is set on various
kinds of export declarations, so that top-level await is available in them.
For example, it enables the following code at the top level of modules:
- export var x = await call();
- export default await call();
- export const y = call();

This patch also includes the editorial change to set the [+Await] parameter
to some productions where it will have no particular effect (because await
is already banned as an identifier inside modules, and this just has the effect
of banning it again), since it seems logically right, e.g., for import
declarations and exported hoistable declarations.